requirements.txt
torch>=2.2
torchvision>=0.17
scipy>=1.11
numpy>=1.26
matplotlib>=3.8
scikit-learn>=1.4
pyyaml>=6.0
wandb>=0.16
# For the intern task (Swin Transformer backbone):
timm>=1.0.9
